Runcommand Plugin: make the runcommand plugin more secure by default
Drop the `--talk=org.freedesktop.Flatpak` permission and default to
executing commands with `flatpak-spawn` when available.

Flatpak users will have to manually poke a hole in the sandbox to
execute commands on the host and all users will have to disable the
`isolate-subprocesses` setting, which comes with a warning.

This is more inconvenient for the common use-case of the plugin, which
typically involves running commands on the host, but it prevents user
plugins from arbitrarily executing commands on the host for flatpak
users.