Plugins: fix user plugin loading in flatpak
When we're on the host we can build a path from `g_get_user_data_dir()`,
but when we're in a flatpak we need to expose `xdg-data/valent/plugins`
and manually build the path from `g_get_home_dir()` or the env will
confuse us.

This also means files in `~/.local/share/valent/plugins` can not be
symbolic links to locations outside of the directory.