ValentDevice: add state property
Add a state property which is a bitmask representing the connected and
paired state of the device, including any transient states.

This seems to be a more reasonable way to expose the state of the device
while allowing for more complex states such as when pairing is in
progress. Since we rarely check the `connected` and `paired` properties
exclusively, this may ultimately replace those as formal properties.