ValentDevice: fix a use-after-free error
When a device emits `notify::state`, it may result in the manager
dropping the last reference if it has become unpaired and disconnected.
Because this would happen in the signal handler, the manager would end
up dropping the last reference to the device leaving it and any other
signal handlers in a use-after-free situation.

When we find ourselves in that situation, remove the device from the
manager but defer dropping the last reference to an idle callback. Also
remove extra emissions of `notify::state` and ensure the rest are the
last operation before the calling function returns.