ValentDevicePanel: fix a possible use-after-free error
If a device is disconnected and the user unpairs it from the device
panel (ie. `device.unpair` action), the device may result in the device
being destroyed before `valent_device_panel_dispose()` is invoked, where
we attempt to disconnect from the device's property changes.

The device manager will let us know when the device has been removed,
so there's no reason for the device panel not to hold a reference.