ValentDevicePlugin: optimize valent_device_plugin_update_state()
Every plugin that implements this interface immediately makes calls to
`valent_device_get_connected()` and `valent_device_get_paired()`, which
both involve locking a mutex now.

Since plugins are always interested in the device state, get that via a
single call to `valent_device_get_state()` and pass it as an argument.