valent-packet: replace macros with convenience function
Replace the `valent_packet_check_*` macros with functions that operate
on the packet `JsonNode`.

The semantics of the functions make more sense and are more convenient
in C where they take a pointer to the field type and return a success
boolean.

These are a bit slower, but safer and make it easier to warn when a
packet is malformed.