packet: use goffset for all payload sizes
Although most of the I/O functions in GIO take and return `gssize`,
`goffset` makes more sense for the payload size in packets since it is
typically describing a file.

In the unlikely scenario that a payload actually exceeds `G_MAXSSIZE`,
it's safer to truncate the read/written size on eg. a splice than it is
to overflow elsewhere.