refactor(ValentDeviceManager): simplify state persistence
The pre-population of known devices is a really a function of the device
manager, not channel services which are actually just responsible for
managing which channel/device is paired.

Move the storage of the device manager "state" to a single JSON file in
the cache directory, which makes it much more obvious when to read and
write the file. The device manager then simply maintains a `JsonNode` of
managed devices, based on their state.