test: propagate output from python wrapper for timeouts
If a test run with the `glibtest.py` wrapper, or a dbusmock derivation,
times out, be sure to propagate the output to aid in debugging.

Unfortunately, if the meson timeout is reached before the script's then
the process is killed and the output never propagated. Make a note of
this if someone goes looking for a reason.