feat(packet): add a `max_len` argument to `valent_packet_from_stream()`
The convenience function `valent_packet_from_stream()` is intended to
read from a stream until reaching a newline character, then attempt to
parse the result as a KDE Connect JSON packet.

In practice, this function is used to read the identity packet from
potentially untrusted peers, so add an optional `max_len` argument to
protect against resource-based Denial-of-Service attacks.