fix: temporary fix for misplaced certificate
After porting to `ValentContext`, plugins store data in their own
context, while Valent still expects the `lan` plugin to look for the
certificate in the root context.

Add a temporary workaround with a call to `valent_context_get_root()`,
which is also safe in tests where there is no parent context.