refactor(device): rework the device manager as an application plugin
The primary difference between the device manager and other components,
turns out to be that it needs to perform certain setup steps in tandem
with the application startup and shutdown phase.

With the `ValentApplicationPlugin` base-class extended to include the
necessary handlers, refactor `ValentDeviceManager` as a subclass
without changing how it's constructed (for now).

This allows absorbing 4 methods unique to `ValentDeviceManager` into the
`ValentApplicationPlugin` base-class.