fix(clipboard): prevent an infinite loop with clipboard content
When content is copied from the remote device to the local clipboard,
the adapter (and then `Valent.Clipboard`) will emit the `::changed`
signal.

This is necessary for other devices to be able to sync the new content
to their remote devices, but must be ignored by the plugin that "owns"
the content or it will be sent back to the source and recurse.

Add an `is_owner` flag to the `ValentClipboardPlugin` `struct` to
indicate that the next `::changed` emission should be ignored by the
plugin.