feat(lan): support DNS-SD
Add support for device discovery with mDNS and DNS-SD.

Currently this is just a way to discover the host and port of a remote
device, which is then used to initiate the standard UDP connection
process. The immediate benefit is avoiding UDP broadcasts, which are
limited by MTU and sometimes dropped entirely if oversized, by sending
the UDP identification packet directly to the device.

The implementation uses the Avahi D-Bus service directly, rather than
add a dependency on `avahi-client` and `avahi-glib` which effectively
do the same thing anyways.