fix(gnome): don't request a ScreenCast session
Sometime between GNOME 44-45 the requirement of a ScreenCast session
for a RemoteDesktop session to work was dropped, although I can't find
the change.

Just drop the request for now, since the reported error is not
specific enough to guess what to do. If necessary, we can probably
query a version somewhere if users of gnome-shell <= 44 report this
breaks for them.

closes #587