fix(sms): avoid a multi-device infinite-loop
Due to some assumptions about the operation of the SMS plugin, there
has been a likely, if not certain, multi-device infinite loop when
requesting messages from the remote device.

Fix the immediate issue by only requesting a thread in the case of a
single message, which should be further restricted by the start date
of the most recent message cached locally.