refactor(mpris): export the player that most recently started playing
When running in a sandbox only one MPRIS player is exported at a time,
which in principle avoids exporting misbehaving or hidden players. In
practice, the current policy sometimes results in a misbehaving player
clobbering the more reasonable choice.

Adopt a policy that favours the player that most recently started
playing, under the assumption that this is the focus of the attention.

closes #638