fix(device): add a temporary fix for device plugins
Due to the way device connections can be "hot swapped", while
incoming packets resolve asynchronously, it's possible for a
device to receive packets while technically disconnected.

Address this by tracking the device state and stashing incoming
packets to prevent violating what should be an invariant that
plugins rely on. Not handling this results in non-deterministic
crashes that are likely more annoying than missing a few packets.