fix(mpris): temporary fix for players not being unexported
When a device disconnects, all of its exported players are dropped
by the adapter, but since there are other references held it
doesn't always result in the player being destroyed or unexported.

This should be fixed properly by implementing a proper object
tree with parental ownership, but it's pretty annoying so implement
a temporary fix.

closes #802